api/requirements.txt
flask==2.2.2
gunicorn==20.1.0
werkzeug==2.3.7
ffmpeg==1.4
openai>=1.24.0,<2.0.0
manim==0.18.0
google-cloud-storage==2.10.0
python-dotenv==1.0.0
bs4==0.0.1
langchain==0.1.16
langchain_community==0.0.34
langchain-openai==0.1.6
langchain-anthropic==0.1.11
langgraph==0.0.38
anthropic==0.37.1
qianfan==0.3.11
google-cloud-logging==3.10.0
flask-cors==3.0.10
azure-storage-blob==12.18.2
azure-monitor-opentelemetry==1.4.2
manim-physics==0.4.0
httpx==0.27.2
google-genai>=0.5.0
